+++ scripts/rollback/rollback-3.2.1.sh
@@ -1,0 +1,221 @@
+#!/bin/bash
+# timestamp: 1648383439899
+
+npm dist-tag add @midwayjs/egg-layer@3.2.0 latest
+npm dist-tag add @midwayjs/egg-layer@3.2.1 beta
+npm dist-tag add @midwayjs/serverless-app@3.2.0 latest
+npm dist-tag add @midwayjs/serverless-app@3.2.1 beta
+npm dist-tag add @midwayjs/axios@3.2.0 latest
+npm dist-tag add @midwayjs/axios@3.2.1 beta
+npm dist-tag add @midwayjs/bootstrap@3.2.0 latest
+npm dist-tag add @midwayjs/bootstrap@3.2.1 beta
+npm dist-tag add @midwayjs/cache@3.2.0 latest
+npm dist-tag add @midwayjs/cache@3.2.1 beta
+npm dist-tag add @midwayjs/consul@3.2.0 latest
+npm dist-tag add @midwayjs/consul@3.2.1 beta
+npm dist-tag add @midwayjs/cos@3.2.0 latest
+npm dist-tag add @midwayjs/cos@3.2.1 beta
+npm dist-tag add @midwayjs/cross-domain@3.2.0 latest
+npm dist-tag add @midwayjs/cross-domain@3.2.1 beta
+npm dist-tag add @midwayjs/express-session@3.2.0 latest
+npm dist-tag add @midwayjs/express-session@3.2.1 beta
+npm dist-tag add @midwayjs/grpc@3.2.0 latest
+npm dist-tag add @midwayjs/grpc@3.2.1 beta
+npm dist-tag add @midwayjs/http-proxy@3.2.0 latest
+npm dist-tag add @midwayjs/http-proxy@3.2.1 beta
+npm dist-tag add @midwayjs/i18n@3.2.0 latest
+npm dist-tag add @midwayjs/i18n@3.2.1 beta
+npm dist-tag add @midwayjs/info@3.2.0 latest
+npm dist-tag add @midwayjs/info@3.2.1 beta
+npm dist-tag add @midwayjs/jwt@3.2.0 latest
+npm dist-tag add @midwayjs/jwt@3.2.1 beta
+npm dist-tag add @midwayjs/mongoose@3.2.0 latest
+npm dist-tag add @midwayjs/mongoose@3.2.1 beta
+npm dist-tag add @midwayjs/orm@3.2.0 latest
+npm dist-tag add @midwayjs/orm@3.2.1 beta
+npm dist-tag add @midwayjs/oss@3.2.0 latest
+npm dist-tag add @midwayjs/oss@3.2.1 beta
+npm dist-tag add @midwayjs/otel@3.2.0 latest
+npm dist-tag add @midwayjs/otel@3.2.1 beta
+npm dist-tag add @midwayjs/passport@3.2.0 latest
+npm dist-tag add @midwayjs/passport@3.2.1 beta
+npm dist-tag add @midwayjs/process-agent@3.2.0 latest
+npm dist-tag add @midwayjs/process-agent@3.2.1 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.2.0 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.2.1 beta
+npm dist-tag add @midwayjs/prometheus@3.2.0 latest
+npm dist-tag add @midwayjs/prometheus@3.2.1 beta
+npm dist-tag add @midwayjs/rabbitmq@3.2.0 latest
+npm dist-tag add @midwayjs/rabbitmq@3.2.1 beta
+npm dist-tag add @midwayjs/redis@3.2.0 latest
+npm dist-tag add @midwayjs/redis@3.2.1 beta
+npm dist-tag add @midwayjs/security@3.2.0 latest
+npm dist-tag add @midwayjs/security@3.2.1 beta
+npm dist-tag add @midwayjs/sequelize@3.2.0 latest
+npm dist-tag add @midwayjs/sequelize@3.2.1 beta
+npm dist-tag add @midwayjs/session@3.2.0 latest
+npm dist-tag add @midwayjs/session@3.2.1 beta
+npm dist-tag add @midwayjs/socketio@3.2.0 latest
+npm dist-tag add @midwayjs/socketio@3.2.1 beta
+npm dist-tag add @midwayjs/static-file@3.2.0 latest
+npm dist-tag add @midwayjs/static-file@3.2.1 beta
+npm dist-tag add @midwayjs/swagger@3.2.0 latest
+npm dist-tag add @midwayjs/swagger@3.2.1 beta
+npm dist-tag add @midwayjs/tablestore@3.2.0 latest
+npm dist-tag add @midwayjs/tablestore@3.2.1 beta
+npm dist-tag add @midwayjs/task@3.2.0 latest
+npm dist-tag add @midwayjs/task@3.2.1 beta
+npm dist-tag add @midwayjs/typegoose@3.2.0 latest
+npm dist-tag add @midwayjs/typegoose@3.2.1 beta
+npm dist-tag add @midwayjs/upload@3.2.0 latest
+npm dist-tag add @midwayjs/upload@3.2.1 beta
+npm dist-tag add @midwayjs/validate@3.2.0 latest
+npm dist-tag add @midwayjs/validate@3.2.1 beta
+npm dist-tag add @midwayjs/version@3.2.0 latest
+npm dist-tag add @midwayjs/version@3.2.1 beta
+npm dist-tag add @midwayjs/view-ejs@3.2.0 latest
+npm dist-tag add @midwayjs/view-ejs@3.2.1 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.2.0 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.2.1 beta
+npm dist-tag add @midwayjs/view@3.2.0 latest
+npm dist-tag add @midwayjs/view@3.2.1 beta
+npm dist-tag add @midwayjs/express@3.2.0 latest
+npm dist-tag add @midwayjs/express@3.2.1 beta
+npm dist-tag add @midwayjs/koa@3.2.0 latest
+npm dist-tag add @midwayjs/koa@3.2.1 beta
+npm dist-tag add @midwayjs/web@3.2.0 latest
+npm dist-tag add @midwayjs/web@3.2.1 beta
+npm dist-tag add @midwayjs/ws@3.2.0 latest
+npm dist-tag add @midwayjs/ws@3.2.1 beta
+tnpm dist-tag add @midwayjs/egg-layer@3.2.0 latest
+tnpm dist-tag add @midwayjs/egg-layer@3.2.1 beta
+tnpm dist-tag add @midwayjs/serverless-app@3.2.0 latest
+tnpm dist-tag add @midwayjs/serverless-app@3.2.1 beta
+tnpm dist-tag add @midwayjs/axios@3.2.0 latest
+tnpm dist-tag add @midwayjs/axios@3.2.1 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.2.0 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.2.1 beta
+tnpm dist-tag add @midwayjs/cache@3.2.0 latest
+tnpm dist-tag add @midwayjs/cache@3.2.1 beta
+tnpm dist-tag add @midwayjs/consul@3.2.0 latest
+tnpm dist-tag add @midwayjs/consul@3.2.1 beta
+tnpm dist-tag add @midwayjs/cos@3.2.0 latest
+tnpm dist-tag add @midwayjs/cos@3.2.1 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.2.0 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.2.1 beta
+tnpm dist-tag add @midwayjs/express-session@3.2.0 latest
+tnpm dist-tag add @midwayjs/express-session@3.2.1 beta
+tnpm dist-tag add @midwayjs/grpc@3.2.0 latest
+tnpm dist-tag add @midwayjs/grpc@3.2.1 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.2.0 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.2.1 beta
+tnpm dist-tag add @midwayjs/i18n@3.2.0 latest
+tnpm dist-tag add @midwayjs/i18n@3.2.1 beta
+tnpm dist-tag add @midwayjs/info@3.2.0 latest
+tnpm dist-tag add @midwayjs/info@3.2.1 beta
+tnpm dist-tag add @midwayjs/jwt@3.2.0 latest
+tnpm dist-tag add @midwayjs/jwt@3.2.1 beta
+tnpm dist-tag add @midwayjs/mongoose@3.2.0 latest
+tnpm dist-tag add @midwayjs/mongoose@3.2.1 beta
+tnpm dist-tag add @midwayjs/orm@3.2.0 latest
+tnpm dist-tag add @midwayjs/orm@3.2.1 beta
+tnpm dist-tag add @midwayjs/oss@3.2.0 latest
+tnpm dist-tag add @midwayjs/oss@3.2.1 beta
+tnpm dist-tag add @midwayjs/otel@3.2.0 latest
+tnpm dist-tag add @midwayjs/otel@3.2.1 beta
+tnpm dist-tag add @midwayjs/passport@3.2.0 latest
+tnpm dist-tag add @midwayjs/passport@3.2.1 beta
+tnpm dist-tag add @midwayjs/process-agent@3.2.0 latest
+tnpm dist-tag add @midwayjs/process-agent@3.2.1 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.2.0 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.2.1 beta
+tnpm dist-tag add @midwayjs/prometheus@3.2.0 latest
+tnpm dist-tag add @midwayjs/prometheus@3.2.1 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.2.0 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.2.1 beta
+tnpm dist-tag add @midwayjs/redis@3.2.0 latest
+tnpm dist-tag add @midwayjs/redis@3.2.1 beta
+tnpm dist-tag add @midwayjs/security@3.2.0 latest
+tnpm dist-tag add @midwayjs/security@3.2.1 beta
+tnpm dist-tag add @midwayjs/sequelize@3.2.0 latest
+tnpm dist-tag add @midwayjs/sequelize@3.2.1 beta
+tnpm dist-tag add @midwayjs/session@3.2.0 latest
+tnpm dist-tag add @midwayjs/session@3.2.1 beta
+tnpm dist-tag add @midwayjs/socketio@3.2.0 latest
+tnpm dist-tag add @midwayjs/socketio@3.2.1 beta
+tnpm dist-tag add @midwayjs/static-file@3.2.0 latest
+tnpm dist-tag add @midwayjs/static-file@3.2.1 beta
+tnpm dist-tag add @midwayjs/swagger@3.2.0 latest
+tnpm dist-tag add @midwayjs/swagger@3.2.1 beta
+tnpm dist-tag add @midwayjs/tablestore@3.2.0 latest
+tnpm dist-tag add @midwayjs/tablestore@3.2.1 beta
+tnpm dist-tag add @midwayjs/task@3.2.0 latest
+tnpm dist-tag add @midwayjs/task@3.2.1 beta
+tnpm dist-tag add @midwayjs/typegoose@3.2.0 latest
+tnpm dist-tag add @midwayjs/typegoose@3.2.1 beta
+tnpm dist-tag add @midwayjs/upload@3.2.0 latest
+tnpm dist-tag add @midwayjs/upload@3.2.1 beta
+tnpm dist-tag add @midwayjs/validate@3.2.0 latest
+tnpm dist-tag add @midwayjs/validate@3.2.1 beta
+tnpm dist-tag add @midwayjs/version@3.2.0 latest
+tnpm dist-tag add @midwayjs/version@3.2.1 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.2.0 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.2.1 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.2.0 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.2.1 beta
+tnpm dist-tag add @midwayjs/view@3.2.0 latest
+tnpm dist-tag add @midwayjs/view@3.2.1 beta
+tnpm dist-tag add @midwayjs/express@3.2.0 latest
+tnpm dist-tag add @midwayjs/express@3.2.1 beta
+tnpm dist-tag add @midwayjs/koa@3.2.0 latest
+tnpm dist-tag add @midwayjs/koa@3.2.1 beta
+tnpm dist-tag add @midwayjs/web@3.2.0 latest
+tnpm dist-tag add @midwayjs/web@3.2.1 beta
+tnpm dist-tag add @midwayjs/ws@3.2.0 latest
+tnpm dist-tag add @midwayjs/ws@3.2.1 beta
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 3.2.0 => 3.2.1
+#  - @midwayjs/serverless-app: 3.2.0 => 3.2.1
+#  - @midwayjs/axios: 3.2.0 => 3.2.1
+#  - @midwayjs/bootstrap: 3.2.0 => 3.2.1
+#  - @midwayjs/cache: 3.2.0 => 3.2.1
+#  - @midwayjs/consul: 3.2.0 => 3.2.1
+#  - @midwayjs/cos: 3.2.0 => 3.2.1
+#  - @midwayjs/cross-domain: 3.2.0 => 3.2.1
+#  - @midwayjs/express-session: 3.2.0 => 3.2.1
+#  - @midwayjs/grpc: 3.2.0 => 3.2.1
+#  - @midwayjs/http-proxy: 3.2.0 => 3.2.1
+#  - @midwayjs/i18n: 3.2.0 => 3.2.1
+#  - @midwayjs/info: 3.2.0 => 3.2.1
+#  - @midwayjs/jwt: 3.2.0 => 3.2.1
+#  - @midwayjs/mongoose: 3.2.0 => 3.2.1
+#  - @midwayjs/orm: 3.2.0 => 3.2.1
+#  - @midwayjs/oss: 3.2.0 => 3.2.1
+#  - @midwayjs/otel: 3.2.0 => 3.2.1
+#  - @midwayjs/passport: 3.2.0 => 3.2.1
+#  - @midwayjs/process-agent: 3.2.0 => 3.2.1
+#  - @midwayjs/prometheus-socket-io: 3.2.0 => 3.2.1
+#  - @midwayjs/prometheus: 3.2.0 => 3.2.1
+#  - @midwayjs/rabbitmq: 3.2.0 => 3.2.1
+#  - @midwayjs/redis: 3.2.0 => 3.2.1
+#  - @midwayjs/security: 3.2.0 => 3.2.1
+#  - @midwayjs/sequelize: 3.2.0 => 3.2.1
+#  - @midwayjs/session: 3.2.0 => 3.2.1
+#  - @midwayjs/socketio: 3.2.0 => 3.2.1
+#  - @midwayjs/static-file: 3.2.0 => 3.2.1
+#  - @midwayjs/swagger: 3.2.0 => 3.2.1
+#  - @midwayjs/tablestore: 3.2.0 => 3.2.1
+#  - @midwayjs/task: 3.2.0 => 3.2.1
+#  - @midwayjs/typegoose: 3.2.0 => 3.2.1
+#  - @midwayjs/upload: 3.2.0 => 3.2.1
+#  - @midwayjs/validate: 3.2.0 => 3.2.1
+#  - @midwayjs/version: 3.2.0 => 3.2.1
+#  - @midwayjs/view-ejs: 3.2.0 => 3.2.1
+#  - @midwayjs/view-nunjucks: 3.2.0 => 3.2.1
+#  - @midwayjs/view: 3.2.0 => 3.2.1
+#  - @midwayjs/express: 3.2.0 => 3.2.1
+#  - @midwayjs/koa: 3.2.0 => 3.2.1
+#  - @midwayjs/web: 3.2.0 => 3.2.1
+#  - @midwayjs/ws: 3.2.0 => 3.2.1