+++ packages/typegoose/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.2.1](https://github.com/midwayjs/midway/compare/v3.2.0...v3.2.1) (2022-03-27)
+
+**Note:** Version bump only for package @midwayjs/typegoose
+
+
+
+
+
 # [3.2.0](https://github.com/midwayjs/midway/compare/v3.1.6...v3.2.0) (2022-03-25)
 
 **Note:** Version bump only for package @midwayjs/typegoose